+++ src/tools/catch_submit.js
@@ -267,11 +267,14 @@ function get_keypress_handler(ignore) {
         }
       }
 
-      /* trigger a "validate" event on the form to be submitted */
+      /* trigger an "implicit_submit" event on the form to be submitted */
       const implicit_event = trigger_event(event.target.form, 'implicit_submit',
-                                           { cancelable: true });
-      implicit_event.trigger = event.target;
-      implicit_event.submittedVia = submit || event.target;
+        {
+          cancelable: true
+        }, {
+          trigger: event.target,
+          submittedVia: submit || event.target,
+        });
       if (implicit_event.defaultPrevented) {
         /* skip the submit, if implicit submit is canceled */
         return;